requirements.txt
tweepy
# requests
python-dotenv
fastapi
uvicorn
